fix bugs

+++ src/main/java/com/fh/controller/app/response/ResBase.java
@@ -2,9 +2,9 @@ package com.fh.controller.app.response;
 import com.fh.util.Const;
 
 public interface ResBase {
-public static	final double version_code =Const.VERSION;
+public static	final Integer version_code =Const.VERSION;
 
-default double getVersion_code() {
+default Integer getVersion_code() {
 	return version_code;
 }
